Se implemento DoctorScheduleRepositoryTests

+++ src/main/java/com/thiagoporfirio/elvivo/domain/repositories/DoctorScheduleRepository.java
@@ -3,5 +3,5 @@ package com.thiagoporfirio.elvivo.domain.repositories;
 import com.thiagoporfirio.elvivo.domain.entities.DoctorScheduleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DoctorScheduleRepository extends JpaRepository<DoctorScheduleEntity, Long>, IDoctorSpecialtyIdExtended<DoctorScheduleEntity> {
+public interface DoctorScheduleRepository extends JpaRepository<DoctorScheduleEntity, Integer>, IDoctorSpecialtyIdExtended<DoctorScheduleEntity> {
 }
